@@ -3,13 +3,15 @@ description: Listen to Phone Call State and provide call back
 version: 0.0.1
 author: Apurv2017 <apurv.hajare@gmail.com>
 homepage: https://github.com/Apurv2017/PhoneCallState
+publish_to: none
 
 dependencies:
   flutter:
     sdk: flutter
 
 environment:
-  sdk: '>=2.0.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
+  flutter: ">=1.20.0 <2.0.6"
 
 # For information on the generic Dart part of this file, see the
 # following page: https://www.dartlang.org/tools/pub/pubspec
